pyproject.toml
[project]
name = "online-bookstore-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi>=0.115.7",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
]
